Add child issue

+++ app/Http/Controllers/ChildrenController.php
@@ -131,7 +131,7 @@ class ChildrenController extends Controller
                 'medicine_detail' => $request->medicine == 'yes' ? $request->medicine_detail : '',
             ]);
             
-            if ($request->medicine == 'yes' && count($request->medicine_dosage) > 0) {
+            if ($request->medicine == 'yes' && isset($request->medicine_dosage) && count($request->medicine_dosage) > 0) {
                 $children->medicine()->createMany($request->medicine_dosage);
             }
 
@@ -247,7 +247,7 @@ class ChildrenController extends Controller
                 'medicine_detail' => $request->medicine == 'yes' ? $request->medicine_detail : '',
             ]);
 
-            if ($request->medicine == 'yes' && count($request->medicine_dosage) > 0) {
+            if ($request->medicine == 'yes' && isset($request->medicine_dosage) && count($request->medicine_dosage) > 0) {
                 $children->medicine()->delete();
                 $children->medicine()->createMany($request->medicine_dosage);
             }
